refactor: organize core folder structure
